packages/providers/contactsprovider:Updating phone number in database
Modified functions insert and update to invoke ComputeNormalizedNumber
to normalize phone numbers
Prior to this change Contact name is not displayed on calling screen.
The name is also not udated in the call logs.
This is a regression of earlier commit 359c3fe7 where
NameNormalizer is used to normalize phone numbers.
As a result the phone number updated in the contacts database
does not match with the look up phone number.

Change-Id: I7b4fc661b4d9e2902a3589fa11cffb6422bf2965
CRs-fixed: 231684