packages/providers/contactsprovider:Updating phone number in database
Modified functions insert and update to invoke ComputeNormalizedNumber
to normalize phone numbers
Prior to this change Contact name is not displayed on calling screen.
The name is also not udated in the call logs.
This is a regression of earlier commit 359c3fe7 where
NameNormalizer is used to normalize phone numbers.
As a result the phone number updated in the contacts database
does not match with the look up phone number.

Gerrit id: I63753211
CRs-fixed: 231684

Change-Id: Ib8df47ea7265f5253ae9b1716ec7810da35e9886