ContactsProviders: catch sqlitedatabase exception
There is no column account_type in view_contacts, so when share
the contacts via baiduyun, it queries the view, a database
exception will be thrown, app crashes. We try to catch the
exception and print the error log.

We try to catch sqlitedatabase exception when querying the
database.

CRs-Fixed: 502055

Change-Id: Ib6c0da9d947d944eb6256077136507fb478c767c