Contacts: Projection map defined is not proper
When we search contacts, it will query Contacts'db with a projection
which is built by that activity and contains an account type column.
However, the projection map of provider does not have this column.

Account type is in table accounts, so the projection map used in provider
is not proper. We should add the columns we need to that projection map.

CRs-Fixed: 503829

Change-Id: I058a8424c19d5fe570f988b944728e8706593b40