ContactsProvider: Searching function in group editting interface is not powerfull
The logic we used to search contacts in group editing interface
is querying all the contacts' display name that match the given
search query, it does not support searching contacts' other info.

Add new functions and class to enable to do the full text search
matching of name info.

CRs-Fixed: 761634

Change-Id: I1b85ffd3519006c3ec64a7d64975d704fd87a9fd