ContactsProvider: Failed to search formatted phone numbers
When searching phone numbers, snippet column may return us
a formatted phone number, such as 'a b c', SQL's like operation
can not match that number since there is no space in the query
string, such as, we can not match 'a b c' with "LIKE '%bc%'".

Remove the ifnull function, we may get null results in snippet
column because the number does not match the query string. Make
sure the sippet is not null to filter unnecessary results.

Change-Id: I1a35aef31796db5adbedf39efa47be5ec923b395
CRs-Fixed: 770178