ContactsProvider2: Can't add vcard as MMS attachment
When add one contact as MMS attachment, the contact's
lookup uri contains ":". When MMS quering the contact
in contact's DB, because of the ":" as multi lookup
uri's separator in contact's DB, so the contact's lookup
uri will be split, this will cause MMS can not find the
contact in contact's DB.

Because of some contact which belong to hotmail account's
lookup uri contains ":", the lookup uri will be split even
only one lookup uri . so use the "&" to replace the ":" as
the multi lookup uri's separator, then only multi lookup uri
will be split.

Change-Id: I9f6acf932b760b57290a53950d0729467365ab51
CRs-Fixed: 848386